Функции get_feedback_web_app_info и get_task_web_app_info вынесены в отдельный файл.
